update raspberries packing

+++ app/models/orderProduct.model.js
@@ -1,6 +1,6 @@
 module.exports = (sequelize, Sequelize) => {
   const OrderProduct = sequelize.define('order_product', {
-    quantity: Sequelize.INTEGER,
+    quantity: Sequelize.DECIMAL,
     packing: Sequelize.STRING,
   }, { timestamps: false });
